% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Belgium.R
\docType{data}
\name{Belgium}
\alias{Belgium}
\title{Data contains Household Consumption (C) and GDP (Y) for Belgium from 1969 to 1998.}
\format{A \code{data.frame} with 30 rows and 3 variables.
\describe{
       \item{\code{Year}}{A tiem series from the 1969 to 1998 to find the year of structural break}
       \item{\code{C}}{Household Consumption}
       \item{\code{Y}}{Gross Domestic Production (GDP)}
       }}
\usage{
data(Belgium)
}
\description{
data for Household Consumption (C) and GDP (Y) for Japan for years 1969 to 1998 for japan to detect the structural breaks with Heteroskedasticity.
}
\examples{
data(Belgium)

}
\author{
\enumerate{
         \item Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Sami Ullah      (\email{samiullahuos@gmail.com})
         \item Gulfam Haider   (\email{haider.gulfam786@gmail.com})
         }
}
\keyword{datasets}
