% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Japan.R
\docType{data}
\name{Japan}
\alias{Japan}
\title{Data contains Household Consumption (C) and GDP (Y) for Japan from 1978 to 2007.}
\format{A \code{data.frame} with 30 rows and 3 variables.
\describe{
       \item{\code{Year}}{A tiem series from the 1978 to 2007 to find the year of structural break}
       \item{\code{C}}{Household Consumption}
       \item{\code{Y}}{Gross Domestic Production (GDP)}
       }}
\usage{
data(Japan)
}
\description{
data for Household Consumption (C) and GDP (Y) for Japan for years 1978 to 2007 for Japan to detect the structural breaks with Heteroskedasticity.
}
\examples{
data(Japan)

}
\author{
\enumerate{
         \item Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Sami Ullah      (\email{samiullahuos@gmail.com})
         \item Gulfam Haider   (\email{haider.gulfam786@gmail.com})
         }
}
\keyword{datasets}
