% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Srilanka.R
\docType{data}
\name{Srilanka}
\alias{Srilanka}
\title{Data contains Household Consumption (C) and GDP (Y) for Sri Lanka from 1978 to 2006.}
\format{A \code{data.frame} with 29 rows and 3 variables.
\describe{
       \item{\code{Year}}{A tiem series from the 1978 to 2006 to find the year of structural break}
       \item{\code{C}}{Household Consumption}
       \item{\code{Y}}{Gross Domestic Production (GDP)}
       }}
\usage{
data(Srilanka)
}
\description{
data for Household Consumption (C) and GDP (Y) for Japan for years 1978 to 2006 for Sri Lanka to detect the structural breaks with Heteroskedasticity.
}
\examples{
data(Srilanka)

}
\author{
\enumerate{
         \item Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Sami Ullah      (\email{samiullahuos@gmail.com})
         \item Gulfam Haider   (\email{haider.gulfam786@gmail.com})
         }
}
\keyword{datasets}
